+++ Sources/ActorCoreBluetooth/BluetoothCentral.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import CoreBluetooth
+internal import ActorCoreBluetoothRuntime
 
 @MainActor
 public final class BluetoothCentral {