+++ Sources/ActorCoreBluetooth/ConnectedPeripheral.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import CoreBluetooth
+internal import ActorCoreBluetoothRuntime
 
 @MainActor
 public final class ConnectedPeripheral {