+++ Sources/ActorCoreBluetooth/SendableWrappers/BluetoothCharacteristic.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import CoreBluetooth
+internal import ActorCoreBluetoothRuntime
 
 /// Sendable wrapper for Bluetooth characteristics
 public struct BluetoothCharacteristic: Sendable {