+++ Sources/ActorCoreBluetooth/SendableWrappers/BluetoothService.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import CoreBluetooth
+internal import ActorCoreBluetoothRuntime
 
 /// Sendable wrapper for Bluetooth services
 public struct BluetoothService: Sendable {