+++ Sources/ActorCoreBluetooth/SendableWrappers/DiscoveredPeripheral.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import CoreBluetooth
+internal import ActorCoreBluetoothRuntime
 
 /// Represents a discovered peripheral that can be connected to
 public struct DiscoveredPeripheral: Sendable {