+++ Sources/ActorCoreBluetoothRuntime/InternalSendable.swift
@@ -1,6 +1,6 @@
 //
 //  InternalSendable.swift
-//  ActorCoreBluetooth
+//  ActorCoreBluetoothRuntime
 //
 //  Created by Konstantin Polin on 8/1/25.
 //  Licensed under the MIT License. See LICENSE file in the project root.
@@ -9,6 +9,8 @@
 import CoreBluetooth
 
 // MARK: - Sendable Extensions
+// These conformances are isolated to the ActorCoreBluetoothRuntime module
+// and do not leak to client code that imports ActorCoreBluetooth.
 extension CBPeripheral: @unchecked @retroactive Sendable {}
 extension CBCentralManager: @unchecked @retroactive Sendable {}
 extension CBUUID: @unchecked @retroactive Sendable {}